requirements.txt
# smart_breaker/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
pydantic==2.10.4
aiosqlite==0.20.0

# ML
torch==2.5.1
numpy==1.26.4
pandas==2.2.3
pyarrow==14.0.2
scipy==1.14.1

# Optimizer
ortools==9.11.4210

# Calendar
icalendar==6.1.0
python-dateutil==2.9.0

# LLM (Groq uses OpenAI-compatible SDK)
openai==1.58.1
elevenlabs==1.15.0

# Utils
python-multipart==0.0.18
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
